telegrambotweb/requirements.txt
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.9.24
Django==4.1.1
django-environ==0.9.0
django-extensions==3.2.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
python-telegram-bot==20.0a4
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.4
tzlocal==4.2
